+++ tests/Components/Models/ThreadTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Components\Models;
 
+use App\Models\Reply;
 use App\Models\Thread;
 use Carbon\Carbon;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
@@ -10,6 +11,39 @@ use Tests\TestCase;
 class ThreadTest extends TestCase
 {
     use DatabaseMigrations;
+
+    /** @test */
+    public function it_can_find_by_slug()
+    {
+        factory(Thread::class)->create(['slug' => 'foo']);
+
+        $this->assertInstanceOf(Thread::class, Thread::findBySlug('foo'));
+    }
+
+    /** @test */
+    public function it_can_give_an_excerpt_of_its_body()
+    {
+        $thread = factory(Thread::class)->make(['body' => 'This is a pretty long text.']);
+
+        $this->assertEquals('This is...', $thread->excerpt(7));
+    }
+
+    /** @test */
+    function we_can_mark_and_unmark_a_reply_as_the_solution()
+    {
+        $thread = factory(Thread::class)->create();
+        $reply = factory(Reply::class)->create(['replyable_id' => $thread->id()]);
+
+        $this->assertFalse($thread->isSolutionReply($reply));
+
+        $thread->markSolution($reply);
+
+        $this->assertTrue($thread->isSolutionReply($reply));
+
+        $thread->unmarkSolution();
+
+        $this->assertFalse($thread->isSolutionReply($reply));
+    }
 
     /** @test */
     function it_can_retrieve_the_latest_threads_in_a_correct_order()
